12510 move reports to use BaseScript
